Spawn a task with the Op builder txn monitoring instead of ExEx (#417)
## 📝 Summary

Closes #412. Using the same scaffolding of the monitoring task I added
support to track the node updates from a canonical stream from the
provider (which is the same primitive used by the exex). Then, I use a
spawned task to run the new monitoring with the stream once the node has
been initialized.

This is a patch until ExEx is fixed or we outsource the monitoring to an
external entity.

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [x] Run `make lint`
* [x] Run `make test`
* [x] Added tests (if applicable)