fix: don't build flashblocks with more gas than block gas limit (#567)
## 📝 Summary

Currently it's possible for the builder to make more Flashblocks than
intended during a block (likely due to delays). If more flashblocks are
made than desired (e.g. 11 flashblocks for a 2s block w/ 200ms
Flashblocks), the 11th flashblock will have a gas limit that is over the
blocks gas limit and will cause invalid blocks on the sequencer.

## 💡 Motivation and Context

If the builder produces a full block that uses more gas than the gas
limit, the local EL client will treat it as invalid.

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [x] Run `make test`
* [x] Added tests (if applicable)