Finish block building process even when cancel request is found (#606)
## 📝 Summary

This PR introduces three changes:
- It refactors some of the common utilities shared among the tests into
a single `TestHarness` struct which spins the test framework with all
the components. Note that it does not update the tests to use this new
utility.
- It fixes an issue with the block builder that would stop the block
building process and not return any block if a cancel request was found.
This happens when an FCU and a getPayload request are called to close to
each other, the getPayload cancels the block building process, and
getPayload waits forever for a block that will never be built. Now, the
block building finishes.
- It adds an integration test to cover this use case with the new
utility.

## 💡 Motivation and Context

<!--- (Optional) Why is this change required? What problem does it
solve? Remove this section if not applicable. -->

---

## ✅ I have completed the following steps:

* [ ] Run `make lint`
* [ ] Run `make test`
* [ ] Added tests (if applicable)